requirements.txt
pyyaml
schedule
urllib3
